Update return types of Fhirstore methods to match FHIR spec (#64)
* Update return types of Fhirstore methods to match FHIR spec

* catch duplicate key error in create

* fix search test

* Handle andle errors in a better way

* docstrings

+++ fhirstore/fhirstore.py
@@ -1,9 +1,14 @@
 import sys
 import logging
-from typing import Union, Dict, List, Optional
+from typing import Union, Dict, Optional
 import json
 
-import elasticsearch
+from elasticsearch import Elasticsearch
+from elasticsearch.exceptions import (
+    NotFoundError as ESNotFoundError,
+    RequestError as ESRequestError,
+    AuthenticationException as ESAuthenticationException,
+)
 import pydantic
 from pymongo import MongoClient, ASCENDING
 from pymongo.errors import DuplicateKeyError
@@ -14,30 +19,23 @@ from fhirpath.enums import FHIR_VERSION
 from fhirpath.search import SearchContext, Search
 from fhir.resources import construct_fhir_element, FHIRAbstractModel
 from fhir.resources.operationoutcome import OperationOutcome
+from fhir.resources.bundle import Bundle
 
 from fhirstore import ARKHN_CODE_SYSTEMS
+from fhirstore.errors import (
+    FHIRStoreError,
+    NotSupportedError,
+    ValidationError,
+    DuplicateError,
+    RequiredError,
+    NotFoundError,
+)
 from fhirstore.search_engine import ElasticSearchEngine
-
-
-class NotFoundError(Exception):
-    """
-    NotFoundError is returned when a resource was not found in the database.
-    """
-
-    pass
-
-
-class BadRequestError(Exception):
-    """
-    BadRequestError is returned when the client request could not be processed.
-    """
-
-    pass
 
 
 class FHIRStore:
     def __init__(
-        self, mongo_client: MongoClient, es_client: elasticsearch.Elasticsearch, db_name: str,
+        self, mongo_client: MongoClient, es_client: Elasticsearch, db_name: str,
     ):
         self.es = es_client
         self.db = mongo_client[db_name]
@@ -57,7 +55,7 @@ class FHIRStore:
         Drops all collections currently in the database.
         """
         if mongo and not es:
-            raise Exception("You also need to drop ES indices when resetting mongo")
+            raise FHIRStoreError("You also need to drop ES indices when resetting mongo")
 
         if mongo:
             for collection in self.resources:
@@ -100,29 +98,23 @@ class FHIRStore:
                 partialFilterExpression={"identifier": {"$exists": True}},
             )
 
-    def validate_resource_type(self, resource_type):
-        if not resource_type:
-            raise BadRequestError("resourceType is missing in resource")
-        elif resource_type not in self.resources:
-            raise NotFoundError(f'unsupported FHIR resource: "{resource_type}"')
-
     def normalize_resource(self, resource: Union[Dict, FHIRAbstractModel]) -> FHIRAbstractModel:
         if isinstance(resource, dict):
             resource_type = resource.get("resourceType")
-            self.validate_resource_type(resource_type)
-            resource = construct_fhir_element(resource_type, resource)
+            if not resource_type:
+                raise RequiredError("resourceType is missing")
+            elif resource_type not in self.resources:
+                raise NotSupportedError(f'unsupported FHIR resource: "{resource_type}"')
+            return construct_fhir_element(resource.get("resourceType"), resource)
+
         elif not isinstance(resource, FHIRAbstractModel):
-            raise BadRequestError(
-                "Provided resource must be of type Union[Dict, FHIRAbstractModel]"
-            )
+            raise FHIRStoreError("Provided resource must be of type Union[Dict, FHIRAbstractModel]")
 
         return resource
 
-    def format_error(self, errors: List[str], severity="error", code="invalid") -> OperationOutcome:
-        issues = [{"severity": severity, "code": code, "diagnostics": err} for err in errors]
-        return OperationOutcome(issue=issues)
-
-    def create(self, resource: Union[Dict, FHIRAbstractModel]):
+    def create(
+        self, resource: Union[Dict, FHIRAbstractModel]
+    ) -> Union[FHIRAbstractModel, OperationOutcome]:
         """
         Creates a resource. The structure of the resource will be checked
         against its json-schema FHIR definition.
@@ -130,23 +122,25 @@ class FHIRStore:
         Args:
             - resource: either a dict with the resource data or a fhir.resources.FHIRAbstractModel
 
-        Returns: The created resource.
+        Returns: The created resource as fhir.resources.FHIRAbstractModel.
         """
         try:
             resource = self.normalize_resource(resource)
         except pydantic.ValidationError as e:
-            return self.format_error(
-                [
-                    f"{err['msg'] or 'Validation error'}: "
-                    f"{e.model.get_resource_type()}.{'.'.join([str(l) for l in err['loc']])}"
-                    for err in e.errors()
-                ]
-            )
+            return ValidationError(e).format()
+        except FHIRStoreError as e:
+            return e.format()
 
-        res = self.db[resource.resource_type].insert_one(json.loads(resource.json()))
-        return {**resource.dict(), "_id": res.inserted_id}
+        try:
+            self.db[resource.resource_type].insert_one(json.loads(resource.json()))
+        except DuplicateKeyError as e:
+            return DuplicateError(
+                f"Resource {resource.resource_type} {resource.id} already exists: {e}"
+            ).format()
 
-    def read(self, resource_type, instance_id):
+        return resource
+
+    def read(self, resource_type, instance_id) -> Union[FHIRAbstractModel, OperationOutcome]:
         """
         Finds a resource given its type and id.
 
@@ -157,14 +151,16 @@ class FHIRStore:
 
         Returns: The found resource.
         """
-        self.validate_resource_type(resource_type)
+        if resource_type not in self.resources:
+            return NotSupportedError(f'unsupported FHIR resource: "{resource_type}"').format()
 
         res = self.db[resource_type].find_one({"id": instance_id}, projection={"_id": False})
         if res is None:
-            raise NotFoundError
-        return res
+            return NotFoundError(f"{resource_type} with id {instance_id} not found").format()
 
-    def update(self, instance_id, resource):
+        return construct_fhir_element(resource_type, res)
+
+    def update(self, instance_id, resource) -> Union[FHIRAbstractModel, OperationOutcome]:
         """
         Update a resource given its type, id and a resource. It applies
         a "replace" operation, therefore the resource will be overriden.
@@ -182,21 +178,23 @@ class FHIRStore:
         try:
             resource = self.normalize_resource(resource)
         except pydantic.ValidationError as e:
-            return self.format_error(
-                [
-                    f"{err['msg'] or 'Validation error'}: "
-                    f"{e.model.get_resource_type()}.{'.'.join([str(l) for l in err['loc']])}"
-                    for err in e.errors()
-                ]
-            )
+            return ValidationError(e).format()
+        except FHIRStoreError as e:
+            return e.format()
+
         update_result = self.db[resource.resource_type].replace_one(
             {"id": instance_id}, json.loads(resource.json()),
         )
         if update_result.matched_count == 0:
-            return self.format_error([f"{resource.resource_type} with id {instance_id} not found"])
-        return update_result
+            return NotFoundError(
+                f"{resource.resource_type} with id {instance_id} not found"
+            ).format()
 
-    def patch(self, resource_type, instance_id, patch):
+        return resource
+
+    def patch(
+        self, resource_type, instance_id, patch
+    ) -> Union[FHIRAbstractModel, OperationOutcome]:
         """
         Update a resource given its type, id and a patch. It applies
         a "patch" operation rather than a "replace", only the fields
@@ -212,15 +210,30 @@ class FHIRStore:
 
         Returns: The updated resource.
         """
-        self.validate_resource_type(resource_type)
+        if resource_type not in self.resources:
+            return NotSupportedError(f'unsupported FHIR resource: "{resource_type}"').format()
 
-        update_result = self.db[resource_type].update_one({"id": instance_id}, {"$set": patch},)
+        res = self.db[resource_type].find_one({"id": instance_id}, projection={"_id": False})
+        if res is None:
+            return NotFoundError(f"{resource_type} with id {instance_id} not found").format()
+
+        patched_resource = {**construct_fhir_element(resource_type, res).dict(), **patch}
+        try:
+            resource = self.normalize_resource(patched_resource)
+        except pydantic.ValidationError as e:
+            return ValidationError(e).format()
+        except FHIRStoreError as e:
+            return e.format()
+
+        update_result = self.db[resource_type].update_one({"id": instance_id}, {"$set": patch})
         if update_result.matched_count == 0:
-            raise NotFoundError
+            return NotFoundError(f"{resource_type} with id {instance_id} not found").format()
 
-        return update_result
+        return resource
 
-    def delete(self, resource_type, instance_id=None, resource_id=None, source_id=None):
+    def delete(
+        self, resource_type, instance_id=None, resource_id=None, source_id=None
+    ) -> OperationOutcome:
         """
         Deletes a resource given its type and id.
 
@@ -231,10 +244,14 @@ class FHIRStore:
 
         Returns: The id of the deleted resource.
         """
-        self.validate_resource_type(resource_type)
+        if resource_type not in self.resources:
+            return NotSupportedError(f'unsupported FHIR resource: "{resource_type}"').format()
 
         if instance_id:
             res = self.db[resource_type].delete_one({"id": instance_id})
+            if res.deleted_count == 0:
+                return NotFoundError(f"{resource_type} with id {instance_id} not found").format()
+
         elif resource_id:
             res = self.db[resource_type].delete_many(
                 {
@@ -246,6 +263,11 @@ class FHIRStore:
                     }
                 }
             )
+            if res.deleted_count == 0:
+                return NotFoundError(
+                    f"{resource_type} with resource_id {resource_id} not found"
+                ).format()
+
         elif source_id:
             res = self.db[resource_type].delete_many(
                 {
@@ -257,17 +279,27 @@ class FHIRStore:
                     }
                 }
             )
+            if res.deleted_count == 0:
+                return NotFoundError(
+                    f"{resource_type} with source_id {source_id} not found"
+                ).format()
+
         else:
-            raise BadRequestError(
-                "one of: 'instance_id', 'resource_id' or 'source_id' are required"
-            )
+            raise FHIRStoreError("one of: 'instance_id', 'resource_id' or 'source_id' are required")
 
-        if res.deleted_count == 0:
-            raise NotFoundError
+        return OperationOutcome(
+            issue=[
+                {
+                    "severity": "information",
+                    "code": "informational",
+                    "diagnostics": f"deleted {res.deleted_count} {resource_type}",
+                }
+            ]
+        )
 
-        return res.deleted_count
-
-    def search(self, resource_type=None, query_string=None, params=None):
+    def search(
+        self, resource_type=None, query_string=None, params=None
+    ) -> Union[Bundle, OperationOutcome]:
         """
         Searchs for params inside a resource.
         Returns a bundle of items, as required by FHIR standards.
@@ -286,62 +318,27 @@ class FHIRStore:
         Returns: A bundle with the results of the search, as required by FHIR
         search standard.
         """
-        if resource_type:
-            self.validate_resource_type(resource_type)
+        if resource_type and resource_type not in self.resources:
+            return NotSupportedError(f'unsupported FHIR resource: "{resource_type}"').format()
 
         search_context = SearchContext(self.search_engine, resource_type)
         fhir_search = Search(search_context, query_string=query_string, params=params)
         try:
             return fhir_search()
-        except elasticsearch.exceptions.NotFoundError as e:
-            return OperationOutcome(
-                issue=[
-                    {
-                        "severity": "error",
-                        "code": "invalid",
-                        "diagnostics": f"{e.info['error']['index']}"
-                        "is not indexed in the database yet.",
-                    }
-                ]
-            )
-        except elasticsearch.exceptions.RequestError as e:
-            return OperationOutcome(
-                issue=[
-                    {
-                        "severity": "error",
-                        "code": "invalid",
-                        "diagnostics": e.info["error"]["root_cause"],
-                    }
-                ]
-            )
-        except elasticsearch.exceptions.AuthenticationException as e:
-            return OperationOutcome(
-                issue=[
-                    {
-                        "severity": "error",
-                        "code": "invalid",
-                        "diagnostics": e.info["error"]["root_cause"],
-                    }
-                ]
-            )
+        except ESNotFoundError as e:
+            return NotFoundError(
+                f"{e.info['error']['index']} is not indexed in the database yet."
+            ).format()
+        except (ESRequestError, ESAuthenticationException) as e:
+            return FHIRStoreError(e.info["error"]["root_cause"]).format()
+        except ESAuthenticationException as e:
+            return FHIRStoreError(e.info["error"]["root_cause"]).format()
         except pydantic.ValidationError as e:
-            issues = []
-            for err in e.errors():
-                issues.append(
-                    {
-                        "severity": "error",
-                        "code": "invalid",
-                        "diagnostics": f"{err['msg']}: "
-                        f"{','.join([f'{e.model.get_resource_type()}.{l}' for l in err['loc']])}",
-                    }
-                )
-            return OperationOutcome(issue=issues)
+            return ValidationError(e).format()
         except fhirpath.exceptions.ValidationError as e:
-            return OperationOutcome(
-                issue=[{"severity": "error", "code": "invalid", "diagnostics": str(e)}]
-            )
+            return ValidationError(str(e)).format()
 
-    def upload_bundle(self, bundle):
+    def upload_bundle(self, bundle) -> Union[None, OperationOutcome]:
         """
         Upload a bundle of resource instances to the store.
 
@@ -349,11 +346,13 @@ class FHIRStore:
             - bundle: the fhir bundle containing the resources.
         """
         if "resourceType" not in bundle or bundle["resourceType"] != "Bundle":
-            raise Exception("input must be a FHIR Bundle resource")
+            return FHIRStoreError(
+                f"input must be a FHIR Bundle resource, got {bundle.get('resourceType')}"
+            ).format()
 
         for entry in bundle["entry"]:
             if "resource" not in entry:
-                raise Exception("Bundle entry is missing a resource.")
+                return RequiredError("Bundle entry is missing a resource.")
 
             try:
                 res = self.create(entry["resource"])
@@ -365,3 +364,5 @@ class FHIRStore:
 
             except DuplicateKeyError as e:
                 logging.warning(f"Document already existed: {e}")
+
+        return None
